Rename module back to forte
