+++ src/set/mod.rs
@@ -4,5 +4,5 @@ pub use pcset::PitchClassSet;
 mod icv;
 pub use icv::*;
 
-pub mod class;
-pub use class::{SetClass, SetClassForm, InversionForm};
+pub mod forte;
+pub use forte::{SetClass, SetClassForm, InversionForm};
